feat: PetFoodQueryService 통합 테스트
