+++ backend/src/test/java/zipgo/petfood/domain/fixture/PetFoodFixture.java
@@ -4,9 +4,11 @@ import static zipgo.petfood.domain.PrimaryIngredients.builder;
 
 import java.util.List;
 import zipgo.brand.domain.Brand;
+import zipgo.petfood.domain.Functionality;
 import zipgo.petfood.domain.HasStandard;
 import zipgo.petfood.domain.Keyword;
 import zipgo.petfood.domain.PetFood;
+import zipgo.petfood.domain.PrimaryIngredients;
 
 public class PetFoodFixture {
 
@@ -26,6 +28,8 @@ public class PetFoodFixture {
                 .imageUrl("https://github.com/woowacourse-teams/2023-zipgo")
                 .purchaseLink("https://avatars.githubusercontent.com/u/94087228?v=4")
                 .keyword(keyword)
+                .functionality(new Functionality(List.of("튼튼,짱짱")))
+                .primaryIngredients(new PrimaryIngredients(List.of("닭고기,쌀", "말미잘")))
                 .hasStandard(HasStandard.builder().build())
                 .brand(brand)
                 .build();
